requirements.txt
biopython
pypdb
gromacs
seaborn
pandas
pyquaternion
rmsd
biostructmap
